+++ src/main/java/nl/lexemmens/podman/BuildMojo.java
@@ -7,11 +7,19 @@ import org.apache.maven.plugins.annotations.LifecyclePhase;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
 
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
 /**
  * BuildMojo for building container images using Podman
  */
 @Mojo(name = "build", defaultPhase = LifecyclePhase.INSTALL)
 public class BuildMojo extends AbstractPodmanMojo {
+
+    private static final Pattern IMAGE_HASH_PATTERN = Pattern.compile("\\b([A-Fa-f0-9]{11,64})\\b");
 
     /**
      * Indicates if building container images should be skipped
@@ -51,8 +59,75 @@ public class BuildMojo extends AbstractPodmanMojo {
     private void buildContainerImage(ImageConfiguration image, ServiceHub hub) throws MojoExecutionException {
         getLog().info("Building container image...");
 
-        String imageHash = hub.getPodmanExecutorService().build(image);
-        image.setImageHash(imageHash);
+        List<String> processOutput = hub.getPodmanExecutorService().build(image);
+
+        // Read the final image hash
+        String finalImageHash = processOutput.get(processOutput.size() - 1);
+        getLog().debug("Determined final image hash as " + finalImageHash);
+        image.setFinalImageHash(finalImageHash);
+
+        if(image.getBuild().isMultistageContainerFile()) {
+            getLog().info("Detected multistage Containerfile...");
+            detemineImageHashes(image, processOutput);
+        }
+    }
+
+    private void detemineImageHashes(ImageConfiguration image, List<String> processOutput) {
+        // Use size -2 as the last line is the image hash of the final image, which we already captured before.
+        Pattern pattern = image.getBuild().getMultistageDockerfileRegex();
+        getLog().debug("Using regular expression: " + pattern);
+        boolean firstOccurrenceFound = false;
+
+        String currentStage = null;
+        for(int i = 0; i <= processOutput.size() - 2; i++) {
+            String currentLine = processOutput.get(i);
+            Matcher matcher = pattern.matcher(currentLine);
+            boolean matches = matcher.find();
+
+            getLog().debug("Processing line: '" + currentLine + "', matches: " + matches);
+
+            if(matches && !firstOccurrenceFound) {
+                // First occurrence
+                firstOccurrenceFound = true;
+                currentStage = matcher.group(3);
+
+                getLog().info("Current stage: " + currentStage);
+            } else if (matches && currentStage != null) {
+                // If we find it again, it means we reached the end of the previous stage. Thus the hash
+                // must be on the previous line.
+                String lineContainingImageHash = processOutput.get(i - 1);
+                extractAndSaveImageHashFromLine(image, currentStage, lineContainingImageHash);
+
+                // Save the current stage
+                currentStage = matcher.group(3);
+            } else if (i == processOutput.size() - 2) {
+                // Handle the last line
+                extractAndSaveImageHashFromLine(image, currentStage, currentLine);
+            }
+        }
+
+
+        getLog().info("Collected hashes: " + image.getImageHashPerStage());
+    }
+
+    private void extractAndSaveImageHashFromLine(ImageConfiguration image, String currentStage, String lineContainingImageHash) {
+        Optional<String> imageHash = retrieveImageHashFromLine(lineContainingImageHash);
+        if (imageHash.isPresent()) {
+            getLog().info("Found image hash " + imageHash.get() + " for stage " + currentStage);
+            image.getImageHashPerStage().put(currentStage, imageHash.get());
+        } else {
+            getLog().warn("Failed to determine image for stage " + currentStage);
+        }
+    }
+
+    private Optional<String> retrieveImageHashFromLine(String line) {
+        String imageHash = null;
+        Matcher matcher = IMAGE_HASH_PATTERN.matcher(line);
+        if(matcher.find()) {
+            imageHash = matcher.group(1);
+        }
+
+        return Optional.ofNullable(imageHash);
     }
 
     private void tagContainerImage(ImageConfiguration image, ServiceHub hub) throws MojoExecutionException {
@@ -66,8 +141,27 @@ public class BuildMojo extends AbstractPodmanMojo {
             return;
         }
 
-        if (image.getImageHash().isPresent()) {
-            String imageHash = image.getImageHash().get();
+        if(image.getBuild().isMultistageContainerFile() && image.isCustomImageNameForMultiStageContainerfile()) {
+            for(Map.Entry<String, String> stageImage : image.getImageHashPerStage().entrySet()) {
+                for(String imageName : image.getImageNamesByStage(stageImage.getKey())) {
+                    String fullImageName = getFullImageNameWithPushRegistry(imageName);
+
+                    getLog().info("Tagging container image " + stageImage.getValue() + " from stage " + stageImage.getKey() + " as " + fullImageName);
+
+                    hub.getPodmanExecutorService().tag(stageImage.getValue(), fullImageName);
+                }
+            }
+        } else if (image.getBuild().isMultistageContainerFile()) {
+            getLog().warn("Missing container names for multistage Containerfile. Falling back to tagging the final container image.");
+            tagFinalImage(image, hub);
+        } else {
+            tagFinalImage(image, hub);
+        }
+    }
+
+    private void tagFinalImage(ImageConfiguration image, ServiceHub hub) throws MojoExecutionException {
+        if (image.getFinalImageHash().isPresent()) {
+            String imageHash = image.getFinalImageHash().get();
             for (String imageNameWithTag : image.getImageNames()) {
                 String fullImageName = getFullImageNameWithPushRegistry(imageNameWithTag);