+++ src/main/java/nl/lexemmens/podman/image/BuildImageConfiguration.java
@@ -12,13 +12,28 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.*;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+import java.util.stream.Collectors;
 
 import static nl.lexemmens.podman.enumeration.ContainerFormat.OCI;
 
 public class BuildImageConfiguration {
 
-    private static final String DEFAULT_DOCKERFILE = "Dockerfile";
+    /**
+     * This is the regular expression to be used to determine a multistage Containerfiles. For now we only support
+     * named stages.
+     */
+    private static final Pattern MULTISTAGE_DOCKERFILE_REGEX = Pattern.compile(".*(FROM\\s.*)([ASas]\\s)([a-zA-Z].*)");
 
+    /**
+     * The default name of the Containerfile to build.
+     */
+    private static final String DEFAULT_CONTAINERFILE = "Containerfile";
+
+    /**
+     * Configures whether caching should be used to build images.
+     */
     @Parameter
     protected boolean noCache;
 
@@ -33,13 +48,13 @@ public class BuildImageConfiguration {
      * Name of the Dockerfile to use. Defaults to Dockerfile
      */
     @Parameter
-    protected String dockerFile;
+    protected String containerFile;
 
     /**
      * Directory containing the Dockerfile
      */
     @Parameter
-    protected File dockerFileDir;
+    protected File containerFileDir;
 
     /**
      * Specify any labels to be applied to the image
@@ -66,6 +81,9 @@ public class BuildImageConfiguration {
     @Parameter
     protected boolean createLatestTag;
 
+    /**
+     * Specifies the format of the Container image to use
+     */
     @Parameter
     protected ContainerFormat format;
 
@@ -73,6 +91,17 @@ public class BuildImageConfiguration {
      * Will be set when this class is validated using the #initAndValidate() method
      */
     private File outputDirectory;
+
+    /**
+     * Will be set to true when the Containerfile is a multistage Containerfile.
+     */
+    private boolean isMultistageContainerFile;
+
+    /**
+     * List of all build stages (only populated in case of multistage Containerfile)
+     */
+    private List<String> stages = new ArrayList<>();
+
 
     /**
      * Constructor
@@ -116,9 +145,9 @@ public class BuildImageConfiguration {
      *
      * @return A {@link File} object referencing the location of the Dockerfile
      */
-    public Path getSourceDockerfile() {
-        Path dockerFileDirPath = Paths.get(dockerFileDir.toURI());
-        return dockerFileDirPath.resolve(dockerFile);
+    public Path getSourceContainerFileDir() {
+        Path dockerFileDirPath = Paths.get(containerFileDir.toURI());
+        return dockerFileDirPath.resolve(containerFile);
     }
 
     /**
@@ -126,8 +155,8 @@ public class BuildImageConfiguration {
      *
      * @return Returns a path to the target Dockerfile
      */
-    public Path getTargetDockerfile() {
-        return Paths.get(outputDirectory.toURI()).resolve(dockerFile);
+    public Path getTargetContainerFile() {
+        return Paths.get(outputDirectory.toURI()).resolve(containerFile);
     }
 
     /**
@@ -159,8 +188,8 @@ public class BuildImageConfiguration {
      *
      * @return The location of the raw Dockerfile as a File.
      */
-    public File getDockerFileDir() {
-        return dockerFileDir;
+    public File getContainerFileDir() {
+        return containerFileDir;
     }
 
     /**
@@ -172,6 +201,30 @@ public class BuildImageConfiguration {
     }
 
     /**
+     * Returns true when the Containerfile is a multistage Containerfile
+     * @return true when a multistage Containerfile is used
+     */
+    public boolean isMultistageContainerFile() {
+        return isMultistageContainerFile;
+    }
+
+    /**
+     * Returns a list of all stages present in the Containerfile
+     * @return a list of all stages.
+     */
+    public List<String> getStages() {
+        return stages;
+    }
+
+    /**
+     * Returns the Pattern that is used to determine if a line matches a multi-stage Containerfile
+     * @return The pattern to determine if a line matches the expected pattern for a multi-stage Containerfile.
+     */
+    public Pattern getMultistageDockerfileRegex() {
+        return MULTISTAGE_DOCKERFILE_REGEX;
+    }
+
+    /**
      * Validates this class by giving all null properties a default value.
      *
      * @param project The MavenProject used to derive some of the default values from.
@@ -179,12 +232,12 @@ public class BuildImageConfiguration {
      * @throws MojoExecutionException In case there is no Dockerfile at the specified source location or the Dockerfile is empty
      */
     public void validate(MavenProject project, Log log) throws MojoExecutionException {
-        if (dockerFile == null) {
-            dockerFile = DEFAULT_DOCKERFILE;
+        if (containerFile == null) {
+            containerFile = DEFAULT_CONTAINERFILE;
         }
 
-        if (dockerFileDir == null) {
-            dockerFileDir = project.getBasedir();
+        if (containerFileDir == null) {
+            containerFileDir = project.getBasedir();
         }
 
         if (labels == null) {
@@ -198,25 +251,48 @@ public class BuildImageConfiguration {
         this.mavenProjectVersion = project.getVersion();
         this.outputDirectory = new File(project.getBuild().getDirectory());
 
-        Path sourceDockerfile = getSourceDockerfile();
-        if (!Files.exists(sourceDockerfile)) {
-            String msg = "No Dockerfile found at " + sourceDockerfile + ". Check your the dockerFileDir and dockerFile parameters in the configuration.";
+        Path sourceContainerFile = getSourceContainerFileDir();
+        if (!Files.exists(sourceContainerFile)) {
+            String msg = "No Containerfile found at " + sourceContainerFile + ". Check your the containerFileDir and containerFile parameters in the configuration.";
             log.error(msg);
             throw new MojoExecutionException(msg);
         }
 
-        if (isDockerfileEmpty(log, sourceDockerfile)) {
-            String msg = "The specified Dockerfile at " + sourceDockerfile + " is empty!";
+        if (isContainerFileEmpty(log, sourceContainerFile)) {
+            String msg = "The specified Containerfile at " + sourceContainerFile + " is empty!";
             log.error(msg);
             throw new MojoExecutionException(msg);
         }
+
+        determineBuildStages(log, sourceContainerFile);
     }
 
-    private boolean isDockerfileEmpty(Log log, Path fullDockerFilePath) throws MojoExecutionException {
+    private boolean isContainerFileEmpty(Log log, Path fullDockerFilePath) throws MojoExecutionException {
         try {
             return 0 == Files.size(fullDockerFilePath);
         } catch (IOException e) {
-            String msg = "Unable to determine if Dockerfile is empty.";
+            String msg = "Unable to determine if Containerfile is empty.";
+            log.error(msg, e);
+            throw new MojoExecutionException(msg, e);
+        }
+    }
+
+    private void determineBuildStages(Log log, Path fullContainerFilePath) throws MojoExecutionException {
+        try {
+            List<String> content = Files.lines(fullContainerFilePath).collect(Collectors.toList());
+            for(String line : content) {
+                Matcher matcher = MULTISTAGE_DOCKERFILE_REGEX.matcher(line);
+                if(matcher.find()) {
+                    isMultistageContainerFile = true;
+
+                    String stage = matcher.group(3);
+
+                    log.debug("Found a stage named: " + stage);
+                    stages.add(stage);
+                }
+            }
+        } catch (IOException e) {
+            String msg = "Unable to determine if Containerfile is a multistage Containerfile.";
             log.error(msg, e);
             throw new MojoExecutionException(msg, e);
         }